Adds call handlers to bh_memmap and registration of extmethod
This commit adds the memmap function to the core.py file.
The operation of the the new bh.memmap( ... ) function call is similar
to the np.memmap call:
http://docs.scipy.org/doc/numpy/reference/generated/numpy.memmap.html

The change in the node VEM now supports the extension method implemented
by bh_memmap. This change is needed since the dynamicly generated VBC
should never hit the VE since it is handled by VEM's only. As such this
VBC will be nop'ed when recieved.