Fixes BH_SYNC issue, this will tricker a read complete file
This is should of cause be avoided, however if the user of Bohrium
really wants to print the whole array are some stupid thing like this it
needs to be read.

This SYNC fix does however ensure that the whole file is read, in a
continues way in one single read operation.